+++ CHANGELOG.md
@@ -1,9 +1,14 @@
+### May 02 2024
+
+- Created a unit test for addSuffix function in aegean.py
+
 ### Apr 30 2024
 
 General
 - Created a addSuffix function in aegean.py to add a suffix to a filename
 
 - Added mpi4py to the dependencies
+
 ### Apr 16 2024
 
 General